requirements.txt
docker-py==1.3.1
docutils==0.12
multiprocessing==2.6.2.1
plumbum==1.5.0
Pygments==2.0.2
pymongo==3.0.3
python-ldap==2.4.20
pytidylib==0.2.4
PyYAML==3.11
requests==2.7.0
rpyc==3.3.0
sh==1.11
simpleldap==0.8
six==1.9.0
web.py==0.37
websocket-client==0.32.0
wheel==0.24.0
